move modal into separate partial and update appearance by adding Modal component and Buttons
