+++ dev/src/blocks/post-chooser/editor.scss
@@ -4,6 +4,6 @@
  * Replace them with your own styles or remove the file completely.
  */
 
-.wp-block-super-fun-static-super-fun-static-block {
+.wp-block-imports-dev-imports-dev-post-chooser {
 	border: 1px dotted #f00;
 }
